+++ guillotina_cms/api/images.py
@@ -1,4 +1,3 @@
-from aiohttp.web import StreamResponse
 from guillotina import app_settings
 from guillotina import configure
 from guillotina.api.files import DownloadFile
@@ -15,63 +14,68 @@ from guillotina.response import Response
 
 
 @configure.service(
-    context=IHasImage, method='GET', permission='guillotina.ViewContent',
-    name='@@images/{field_name}',
-    **_traversed_file_doc('Download the image'))
+    context=IHasImage,
+    method="GET",
+    permission="guillotina.ViewContent",
+    name="@@images/{field_name}",
+    **_traversed_file_doc("Download the image"),
+)
 class DownloadImageFile(DownloadFile):
     pass
 
 
 @configure.service(
-    context=IHasImage, method='GET', permission='guillotina.ViewContent',
-    name='@@images/{field_name}/{scale}',
-    **_traversed_file_doc('Download the image scale'))
+    context=IHasImage,
+    method="GET",
+    permission="guillotina.ViewContent",
+    name="@@images/{field_name}/{scale}",
+    **_traversed_file_doc("Download the image scale"),
+)
 class DownloadImageScale(TraversableFieldService):
-
     async def __call__(self):
         registry = await get_registry()
         settings = registry.for_interface(IImagingSettings)
-        scale_name = self.request.matchdict['scale']
-        allowed_sizes = settings['allowed_sizes']
+        scale_name = self.request.matchdict["scale"]
+        allowed_sizes = settings["allowed_sizes"]
         if scale_name not in allowed_sizes:
-            raise HTTPNotFound(content={
-                'reason': f'{scale_name} is not supported'
-            })
+            raise HTTPNotFound(content={"reason": f"{scale_name} is not supported"})
         file = self.field.get(self.field.context or self.context)
         if file is None:
-            raise HTTPNotFound(content={
-                'message': 'File or custom filename required to download'
-            })
+            raise HTTPNotFound(
+                content={"message": "File or custom filename required to download"}
+            )
 
         adapter = get_multi_adapter(
-            (self.context, self.request, self.field), IFileManager)
-        data = b''
+            (self.context, self.request, self.field), IFileManager
+        )
+        data = b""
         async for chunk in adapter.iter_data():
             data += chunk
 
-        width, _, height = allowed_sizes[scale_name].partition(':')
+        width, _, height = allowed_sizes[scale_name].partition(":")
 
         result, format_, size = scaleImage(
-            data, int(width), int(height),
-            quality=settings['quality'],
-            direction='thumbnail')
+            data,
+            int(width),
+            int(height),
+            quality=settings["quality"],
+            direction="thumbnail",
+        )
 
-        cors_renderer = app_settings['cors_renderer'](self.request)
+        cors_renderer = app_settings["cors_renderer"](self.request)
         headers = await cors_renderer.get_headers()
-        headers.update({
-            'CONTENT-DISPOSITION': 'attachment; filename="{}"'.format(
-                file.filename)
-        })
+        headers.update(
+            {"CONTENT-DISPOSITION": 'attachment; filename="{}"'.format(file.filename)}
+        )
 
         download_resp = Response(
             status=200,
             headers=headers,
             content_type=f"image/{format_}",
-            content_length=len(result)
+            content_length=len(result),
         )
         await download_resp.prepare(self.request)
 
         await download_resp.write(result)
         await download_resp.write(eof=True)
         return download_resp
-